Position NavBottom with "fixed" instead of "sticky"
This fixes a Firefox Android bug in which the bottom bar is rendered
offset from the actual button location.

Apparently there's a browser bug with bottom-hugging sticky elements.
When the browser's nav bar hides, the sticky element looks like it's
positioned correctly, but the pointer events pass through it.

But replacing `position: sticky` with `position: fixed` is an easy fix.